stream: fix array out-of-bounds in stream_get_timing_info_callback
This issue was found by enabling ubsan. For me it consistently triggered
after about 28 seconds running a simple example that plays a sine wave
via the mainloop api.

I added a log and confirmed that before the ubsan is triggered the
index variable j is indeed 32 which is out-of-bounds.

This commit was cherry-picked from upstream:

https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/commit/1fbc6b0e0e796ea1e78825fce4ca815475d30294

Co-authored-by: Arun Raghavan <arun@asymptotic.io>